Create a demo/POC for the Conforma VSA service
Add some resources that can be used for a demo, (following the
convention set by the existing demo resources).

Create a bash script to show some pod logs, inspect created VSAs
etc, for demo purposes.

(This is not meant to be merged, it's just a reference to help with
rolling this out to Red Hat's internal Konflux instances.)

Ref: https://issues.redhat.com/browse/EC-1528